move monsters into own directory
